// apps/web/src/api-client/model/reportType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */

export type ReportType = (typeof ReportType)[keyof typeof ReportType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportType = {
    INAPPROPRIATE_BEHAVIOR: "INAPPROPRIATE_BEHAVIOR",
    NO_SHOW: "NO_SHOW",
    OVERCHARGING: "OVERCHARGING",
    LEFT_LUGGAGE: "LEFT_LUGGAGE",
    SAFETY_ISSUE: "SAFETY_ISSUE",
    OTHER: "OTHER",
} as const;